requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
drf-spectacular>=0.27
psycopg[binary]>=3.2
python-decouple>=3.8
django-filter>=24.0
django-extensions>=3.2
Pillow
